Добавление сохранение в файл и включение суперсэмплирования

+++ src/main/java/edu/project4/Etinties/Parameters.java
@@ -1,4 +1,4 @@
 package edu.project4.Etinties;
 
-public record Parameters(int param1, int param2, int param3, int param4) {
+public record Parameters(double param1, double param2, double param3, double param4) {
 }
